requirements.txt
Flask==2.2.2
numpy==1.23.3
mediapipe==0.8.11
opencv_python==4.6.0.66
Pillow==9.2.0
pyserial==3.5
